## Describe the changes this PR makes. Why should it be merged?

Adds more lints and then resolves them:
- [x] Adds must_use to constructors and getters
- [x] Documents losses in precision and uses `u32::from` and equivalents
to prevent accidental changes that will cause losses in precision
- [x] Adds semicolons to places that were missing semicolons where it
makes sense to add them
- [x] Documents the errors returned by functions that return `Result`s
   - [x] ADI devices
   - [x] Smart devices
   - [x] Controller
   - [x] Screen
- [x] Adds separators to RGB hexes to improve readability
- [x] `BannerTheme` is now `Copy`
- [x] A few other minor changes suggested by Clippy
- [x] Fixes a bug in which `port.validate_expander()` was not called
- [x] Changes instances of "Gets x" terminology to "Returns x" in getter
docs.
- [x] (breaking) `OpticalSensor::led_brightness` is now returned as a
number from `[0.0, 1.0]` to match `OpticalSensor::set_led_brightness`.

## Additional Context
<!--
Move all applicable items out of the comment:
- I have tested these changes on a VEX V5 brain.
- I have tested these changes in a simulator.
- These are breaking changes (semver: major).
- These are *only* non-code changes (e.g. documentation, README.md).
- These changes update the crate's interface (e.g. functions/modules
added or changed).
-->

+++ packages/vexide-devices/src/smart/distance.rs
@@ -9,7 +9,8 @@ use vex_sdk::{
 use super::{SmartDevice, SmartDeviceType, SmartPort};
 use crate::PortError;
 
-/// A physical distance sensor plugged into a port.
+/// A distance sensor plugged into a smart port.
+///
 /// Distance sensors can only keep track of one object at a time.
 #[derive(Debug, Eq, PartialEq)]
 pub struct DistanceSensor {
@@ -23,7 +24,8 @@ unsafe impl Send for DistanceSensor {}
 unsafe impl Sync for DistanceSensor {}
 
 impl DistanceSensor {
-    /// Create a new distance sensor from a smart port index.
+    /// Creates a new distance sensor from a [`SmartPort`].
+    #[must_use]
     pub fn new(port: SmartPort) -> Self {
         Self {
             device: unsafe { port.device_handle() },
@@ -46,7 +48,13 @@ impl DistanceSensor {
         }
     }
 
-    /// Attempt to detect an object, returning `None` if no object could be found.
+    /// Attempts to detect an object, returning `None` if no object could be found.
+    ///
+    /// # Errors
+    ///
+    /// - A [`DistanceError::Port`] error is returned if there is not a distance sensor connected to the port.
+    /// - A [`DistanceError::StillInitializing`] error is returned if the distance sensor is still initializing.
+    /// - A [`DistanceError::BadStatusCode`] error is returned if the distance sensor has an unknown status code.
     pub fn object(&self) -> Result<Option<DistanceObject>, DistanceError> {
         self.validate()?;
 
@@ -59,13 +67,17 @@ impl DistanceSensor {
                 relative_size: unsafe { vexDeviceDistanceObjectSizeGet(self.device) as u32 },
                 velocity: unsafe { vexDeviceDistanceObjectVelocityGet(self.device) },
                 // TODO: determine if confidence reading is separate from whether or not an object is detected.
-                confidence: unsafe { vexDeviceDistanceConfidenceGet(self.device) as u32 } as f64
+                confidence: f64::from(unsafe { vexDeviceDistanceConfidenceGet(self.device) })
                     / 63.0,
             })),
         }
     }
 
-    /// Gets the status code of the distance sensor
+    /// Returns the internal status code of the distance sensor.
+    ///
+    /// # Errors
+    ///
+    /// - A [`DistanceError::Port`] error is returned if there is not a distance sensor connected to the port.
     pub fn status(&self) -> Result<u32, DistanceError> {
         self.validate_port()?;
 
@@ -88,7 +100,7 @@ impl From<DistanceSensor> for SmartPort {
     }
 }
 
-/// Readings from a phyiscal object detected by a Distance Sensor.
+/// Readings from a physical object detected by a Distance Sensor.
 #[derive(Default, Debug, Clone, PartialEq, PartialOrd)]
 pub struct DistanceObject {
     /// The distance of the object from the sensor (in millimeters).
@@ -122,7 +134,7 @@ pub enum DistanceError {
     /// Need to wait for the sensor to finish initializing
     StillInitializing,
 
-    /// The sensor's status code is not 0x82 or 0x86.
+    /// The sensor has an unknown status code.
     BadStatusCode,
 
     /// Generic port related error.
